# fix: re-enable (xor) bloom filter index