--- a/src/query/service/Cargo.toml
+++ b/src/query/service/Cargo.toml
@@ -89,7 +89,6 @@ backoff = { version = "0.4.0", features = ["futures", "tokio"] }
 backon = "0.2.0"
 base64 = "0.13.0"
 bincode = "2.0.0-rc.1"
-bit-vec = { version = "0.6.3", features = ["serde_std"] }
 bumpalo = "3.11.0"
 byteorder = "1.4.3"
 bytes = "1.2.1"