--- a/src/query/service/tests/it/storages/index/mod.rs
+++ b/src/query/service/tests/it/storages/index/mod.rs
@@ -12,5 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod bloom_filter;
 mod range_filter;