--- a/src/query/storages/fuse/src/pruning/bloom_pruner.rs
+++ b/src/query/storages/fuse/src/pruning/bloom_pruner.rs
@@ -171,11 +171,12 @@ mod util {
             Ok(bloom_filter_index) => BloomFilterIndexer::from_bloom_block(
                 schema.clone(),
                 bloom_filter_index.into_data(),
-                ctx,
             )?
             .maybe_true(filter_expr),
             Err(e) if e.code() == ErrorCode::deprecated_index_format_code() => {
-                // in case that the index is no longer supported, just keep it
+                // In case that the index is no longer supported, just return ture to indicate
+                // that the block being pruned should be kept. (Although the caller of this method
+                // "BloomFilterIndexPruner::should_keep",  will ignore any exceptions returned)
                 Ok(true)
             }
             Err(e) => Err(e),