--- a/src/query/storages/index/Cargo.toml
+++ b/src/query/storages/index/Cargo.toml
@@ -21,10 +21,10 @@ common-fuse-meta = { path = "../fuse-meta" }
 common-legacy-expression = { path = "../../legacy-expression" }
 common-pipeline-transforms = { path = "../../pipeline/transforms" }
 
-bincode = { version = "2.0.0-rc.1", features = ["serde", "std", "alloc"] }
-bit-vec = { version = "0.6.3", features = ["serde_std"] }
 cbordata = { version = "0.6.0" }
-rand = "0.8.5"
 serde = { version = "1.0.145", features = ["derive"] }
 tracing = "0.1.36"
 xorfilter-rs = { git = "https://github.com/datafuse-extras/xorfilter", features = ["cbordata"] }
+
+[dev-dependencies]
+rand = "0.8.5"