--- a/src/query/storages/index/src/bloom/xor8.rs
+++ b/src/query/storages/index/src/bloom/xor8.rs
@@ -17,6 +17,10 @@ use std::hash::Hash;
 use cbordata::Cbor;
 use cbordata::FromCbor;
 use cbordata::IntoCbor;
+use common_datavalues::prelude::TypeID;
+use common_datavalues::remove_nullable;
+use common_datavalues::DataType;
+use common_datavalues::DataTypeImpl;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use xorfilter::Xor8;
@@ -92,4 +96,12 @@ impl Bloom for XorBloom {
         Ok((Self { filter: xor_value }, n))
     }
 }
-impl SupportedType for XorBloom {}
+
+impl SupportedType for XorBloom {
+    fn is_supported_type(data_type: &DataTypeImpl) -> bool {
+        // Bloom index only enabled for String and Boolean types for now
+        let inner_type = remove_nullable(data_type);
+        let data_type_id = inner_type.data_type_id();
+        matches!(data_type_id, TypeID::String)
+    }
+}