--- a/src/query/storages/index/src/lib.rs
+++ b/src/query/storages/index/src/lib.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![deny(unused_crate_dependencies)]
-
 use common_datavalues::prelude::TypeID;
 use common_datavalues::DataType;
 use common_datavalues::DataTypeImpl;