--- a/src/query/storages/index/tests/it/bloom/mod.rs
+++ b/src/query/storages/index/tests/it/bloom/mod.rs
@@ -12,4 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod bloom_filter;
 mod xor;