I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

as xor bloom has been merged in PR https://github.com/datafuselabs/databend/pull/7860, replace the vanilla bloom filter with it.

- for legacy index data of v1 bloom filter, the index will be ignored
  but will be removed if table data is dropped
- xor index is created under path "_i_b_v2"
  so that, other types of indexes could be put in separated "dir"s
- bloom filter index is enabled for string type column only


Fixes #issue