@BohuTANG 2022-09-25
Cool.
The bloom filter should have a type so that we can try fuse binary bloom in the future:
[Binary Fuse Filters: Fast and Smaller Than Xor Filters](https://arxiv.org/pdf/2201.01174.pdf)
![image](https://user-images.githubusercontent.com/172204/192122821-f22ecf85-98d5-4d01-b660-ef5e01fbe8cb.png)