@BohuTANG 2022-09-26
I will bench the xor bloom index loading after https://github.com/datafuselabs/databend/pull/7867 and https://github.com/datafuselabs/databend/pull/7857